refactor account update dialog
