+++ ui/src/features/Dialog/components/settings/index.ts
@@ -1,4 +1,3 @@
-export * from "./AccountDialog";
 export * from "./GeneralDialog";
 export * from "./WorkflowDialog";
 export * from "./WorkspacesDialog";
